+++ esbuild.js
@@ -1,3 +1,0 @@
-import {buildScripts} from '../yii2-skeleton/esbuild.config.js';
-
-await buildScripts();